NoGfx: Implement GpuRasterDesc in pipeline creation.

+++ NoGfx/test/shaders/compute_constants.metal
@@ -2,9 +2,8 @@
 
 using namespace metal;
 
-constant float computeScale [[function_constant(0)]];
+constant uint computeScale [[function_constant(0)]];
 
 [[host_name("main")]] kernel void computeMain() {
-	if (computeScale > 0.0f) {
-	}
+	if (computeScale > 0) {}
 }
