+++ NoGfx/test/shaders/render_fragment_constants.metal
@@ -2,8 +2,8 @@
 
 using namespace metal;
 
-constant float fragmentScale [[function_constant(0)]];
+constant uint fragmentScale [[function_constant(0)]];
 
 [[host_name("main")]] fragment float4 fragmentMain() {
-	return float4(fragmentScale, 0.0, 0.0, 1.0);
+	return float4(float(fragmentScale), 0.0, 0.0, 1.0);
 }
